Browsing a list of users
You were browsing a list of users registered on your site when
you noticed that some email addresses might be duplicated.

Every address consists of a local name and a domain name, separated by the
@ sign. For example, in the address "duplicate@example.com", "duplicate"
is a local name and "example.com" is a domain name.

For the purposes of this task, let's define two ways of formatting an
email address:

1. if you add dots('.') between some characters in the local name
parth of an email address, it will be forwarded to the equivalent
address without dots. For instance, email addresses of the form
"dupli.cate@example.com", "d.u.p.l.c.a.t.e@example.com" will be forwarded
to "duplicate@example.com". Please note that this doesn't work for domain
names, so "duplicate@exa.mple.com.net" is different from "duplicate@example.com.net";

2. if you add a plus("+") in the local name, everything after the plus sign
will be ignored. This allows certain emails to be filtered.

for example:

"duplicate+work@example.com" will be forwarded to "duplicate@example.com"

It is posible to use both of these formatting methods at same time.

You are give a list of addresses. We collect them together into groups,
with each group containing emails that equivalent according to the above
rules.

write fuction that given a list of emails L, will return the number of
groups whose length is greater than one.