Collect Fruit Problem
In a grader there are N fruit trees growing in a row. Each tree produces one
Fruit of some type.

Little Amy goes to the grader to collect some fruit. She has two baskets that
Can hold any number of fruit, but she only wants to put one type of fruit in
Each basket. When Amy enters the grader, she goes to some tree,  collects
The fruit from it (every tree produces exactly one fruit) and then goes to the
Next tree to the right. Amy continues her walk until she encounters a tree with
A fruit that she cannot put in any of her baskets (each of them already has
A fruit of a different type), or when the garden ends. Amy never throws any fruit
Away from her basket.

What is the maximum number of fruit she can collect?

Write a function: that given a array A of N integers, representing consecutive
Types of fruit in the garden, returns the maximal number of fruit Amy can collect.

For example, given A = [1, 2, 1, 3, 4, 3, 5, 1, 2] the function should return 3. Amy
can collect fruit of types 1 and 2 from the beginning of the garden or all the fruits
of types 3 and 4